+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/runner/ForAll2Test.kt
@@ -1,11 +1,13 @@
 package com.github.jcornaz.kwik.runner
 
+import com.github.jcornaz.kwik.FalsifiedPropertyError
 import com.github.jcornaz.kwik.Generator
 import com.github.jcornaz.kwik.forAll
 import com.github.jcornaz.kwik.runner.AbstractRunnerTest
 import com.github.jcornaz.kwik.withSamples
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
 class ForAll2Test : AbstractRunnerTest() {
@@ -35,6 +37,28 @@ class ForAll2Test : AbstractRunnerTest() {
         assertTrue(100 in ints)
         assertTrue(123.0 in doubles)
         assertTrue(123.0 in doubles)
+    }
+
+    @Test
+    fun falsificationDisplayHelpfulMessage() {
+        val exception = assertFailsWith<FalsifiedPropertyError> {
+            var i = 0
+            forAll(
+                Generator.create { 42 },
+                Generator.create { -4.1 },
+                iterations = 123, seed = 78
+            ) { _, _ -> ++i < 12 }
+        }
+
+        assertEquals(
+            """
+                Property falsified after 12 tests (out of 123)
+                Argument 1: 42
+                Argument 2: -4.1
+                Generation seed: 78
+            """.trimIndent(),
+            exception.message
+        )
     }
 
     @Test