+++ CHANGELOG.rst
@@ -9,6 +9,15 @@ and this project adheres to `Semantic Versioning`_.
 .. _Keep a Changelog: https://keepachangelog.com/en/1.0.0
 .. _Semantic Versioning: https://semver.org/spec/v2.0.0.html
 
+Unreleased_
+-----------
+
+Changed
+.......
+
+* Show test details in console only in case of success
+* Improved exception message in case of falsified property (and introduce typed exception ``FalsifiedPropertyError``
+
 0.1.0-alpha.4_ - 2019-06-10
 ---------------------------